# frontend/openapi.json
{
  "components": {
    "schemas": {
      "Assessment": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "errors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/NodeError"
            },
            "title": "Errors",
            "type": "array"
          },
          "fitness": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FitnessResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "injury": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InjuryResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Metrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "mode": {
            "enum": [
              "mock",
              "live"
            ],
            "title": "Mode",
            "type": "string"
          },
          "normalized": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NormalizedInputs"
              },
              {
                "type": "null"
              }
            ]
          },
          "nutrition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NutritionPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileIn"
          },
          "progress_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProgressContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "recovery": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecoveryPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "safety": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SafetyResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "sections": {
            "additionalProperties": {
              "enum": [
                "ok",
                "partial",
                "failed",
                "skipped"
              ],
              "type": "string"
            },
            "default": {},
            "title": "Sections",
            "type": "object"
          },
          "simulated": {
            "description": "True when produced by the mock generator rather than a live model",
            "title": "Simulated",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "complete",
              "partial",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "warnings": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          },
          "workout": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkoutPlan"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "created_at",
          "mode",
          "simulated",
          "status",
          "profile"
        ],
        "title": "Assessment",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field_errors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Field Errors",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FieldError": {
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
      "FitnessResult": {
        "properties": {
          "confidence": {
            "description": "Probability of the predicted class, in percent (0-100).",
            "title": "Confidence",
            "type": "number"
          },
          "drivers": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Driver"
            },
            "title": "Drivers",
            "type": "array"
          },
          "level": {
            "title": "Level",
            "type": "string"
          },
          "notes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "probabilities": {
            "additionalProperties": {
              "type": "number"
            },
            "description": "Probability of each class as a fraction (0-1); they sum to 1.",
            "title": "Probabilities",
            "type": "object"
          }
        },
        "required": [
          "level",
          "confidence",
          "probabilities"
        ],
        "title": "FitnessResult",
        "type": "object"
      },
...
      "InjuryResult": {
        "properties": {
          "confidence": {
            "description": "Probability of the predicted class, in percent (0-100).",
            "title": "Confidence",
            "type": "number"
          },
          "drivers": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Driver"
            },
            "title": "Drivers",
            "type": "array"
          },
          "notes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "probabilities": {
            "additionalProperties": {
              "type": "number"
            },
            "description": "Probability of each class as a fraction (0-1); they sum to 1.",
            "title": "Probabilities",
            "type": "object"
          },
          "risk": {
            "title": "Risk",
            "type": "string"
          },
          "risk_factors": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Risk Factors",
            "type": "array"
          }
        },
        "required": [
          "risk",
          "confidence",
          "probabilities"
        ],
        "title": "InjuryResult",
        "type": "object"
      },
...
      "Metrics": {
        "properties": {
          "age_category": {
            "title": "Age Category",
            "type": "string"
          },
          "bmi": {
            "title": "Bmi",
            "type": "number"
          },
          "bmi_category": {
            "title": "Bmi Category",
            "type": "string"
          }
        },
        "required": [
          "bmi",
          "bmi_category",
          "age_category"
        ],
        "title": "Metrics",
        "type": "object"
      },
      "NodeError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "node": {
            "title": "Node",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "node",
          "code",
          "message"
        ],
        "title": "NodeError",
        "type": "object"
      },
...
      "NormalizedInputs": {
        "properties": {
          "experience": {
            "$ref": "#/components/schemas/NormalizedExperience"
          },
          "health": {
            "$ref": "#/components/schemas/NormalizedHealth"
          },
          "schedule": {
            "$ref": "#/components/schemas/NormalizedSchedule"
          }
        },
        "required": [
          "experience",
          "health",
          "schedule"
        ],
        "title": "NormalizedInputs",
        "type": "object"
      },
...
      "ProgressLogIn": {
        "additionalProperties": false,
        "properties": {
          "completed": {
            "default": true,
            "title": "Completed",
            "type": "boolean"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "maximum": 600.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "kind": {
            "enum": [
              "session",
              "weight"
            ],
            "title": "Kind",
            "type": "string"
          },
          "logged_on": {
            "format": "date",
            "title": "Logged On",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "planned_day": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Day"
          },
          "rpe": {
            "anyOf": [
              {
                "maximum": 10.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Rate of perceived exertion, 1-10",
            "title": "Rpe"
          },
          "weight_kg": {
            "anyOf": [
              {
                "maximum": 300.0,
                "minimum": 30.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          }
        },
        "required": [
          "kind"
        ],
        "title": "ProgressLogIn",
        "type": "object"
      },
      "ProgressLogOut": {
        "additionalProperties": false,
        "properties": {
          "assessment_id": {
            "title": "Assessment Id",
            "type": "string"
          },
          "completed": {
            "default": true,
            "title": "Completed",
            "type": "boolean"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "maximum": 600.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kind": {
            "enum": [
              "session",
              "weight"
            ],
            "title": "Kind",
            "type": "string"
          },
          "logged_on": {
            "format": "date",
            "title": "Logged On",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "planned_day": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Day"
          },
          "rpe": {
            "anyOf": [
              {
                "maximum": 10.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Rate of perceived exertion, 1-10",
            "title": "Rpe"
          },
          "weight_kg": {
            "anyOf": [
              {
                "maximum": 300.0,
                "minimum": 30.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          }
        },
        "required": [
          "kind",
          "id",
          "assessment_id"
        ],
        "title": "ProgressLogOut",
        "type": "object"
      },
      "ProgressSummary": {
        "properties": {
          "adherence_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adherence Pct"
          },
          "assessment_id": {
            "title": "Assessment Id",
            "type": "string"
          },
          "average_rpe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Rpe"
          },
          "deload_due": {
            "title": "Deload Due",
            "type": "boolean"
          },
          "latest_weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Weight Kg"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "planned_sessions_per_week": {
            "title": "Planned Sessions Per Week",
            "type": "integer"
          },
          "sessions_last_28_days": {
            "title": "Sessions Last 28 Days",
            "type": "integer"
          },
          "weekly": {
            "items": {
              "$ref": "#/components/schemas/WeekAdherence"
            },
            "title": "Weekly",
            "type": "array"
          },
          "weeks_on_plan": {
            "title": "Weeks On Plan",
            "type": "number"
          },
          "weight_change_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Change Kg"
          },
          "weights": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array"
            },
            "title": "Weights",
            "type": "array"
          }
        },
        "required": [
          "assessment_id",
          "planned_sessions_per_week",
          "sessions_last_28_days",
          "adherence_pct",
          "average_rpe",
          "latest_weight_kg",
          "weight_change_kg",
          "weeks_on_plan",
          "deload_due",
          "weekly",
          "weights",
          "notes"
        ],
        "title": "ProgressSummary",
        "type": "object"
      },
...
      "WeekAdherence": {
        "properties": {
          "planned": {
            "title": "Planned",
            "type": "integer"
          },
          "sessions": {
            "title": "Sessions",
            "type": "integer"
          },
          "week_start": {
            "format": "date",
            "title": "Week Start",
            "type": "string"
          }
        },
        "required": [
          "week_start",
          "sessions",
          "planned"
        ],
        "title": "WeekAdherence",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Personalised fitness assessments combining calibrated ML classifiers with an LLM planning workflow. Live generation is bring-your-own-key: send your Gemini key in `X-Gemini-Api-Key` (never stored or logged).",
    "title": "AetherFit AI API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/assessments/{assessment_id}": {
      "delete": {
        "operationId": "delete_assessment_api_v1_assessments__assessment_id__delete",
        "parameters": [
          {
            "description": "Assessment UUID",
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "description": "Assessment UUID",
              "maxLength": 36,
              "minLength": 36,
              "title": "Assessment Id",
              "type": "string"
            }
          },
          {
            "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
            "in": "header",
            "name": "x-device-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
              "title": "X-Device-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid or missing Gemini API key (live mode)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Rate limited (this API or the Gemini key's quota)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Delete Assessment",
        "tags": [
          "assessments"
        ]
      },
      "get": {
        "operationId": "get_assessment_api_v1_assessments__assessment_id__get",
        "parameters": [
          {
            "description": "Assessment UUID",
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "description": "Assessment UUID",
              "maxLength": 36,
              "minLength": 36,
              "title": "Assessment Id",
              "type": "string"
            }
          },
          {
            "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
            "in": "header",
            "name": "x-device-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
              "title": "X-Device-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assessment"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid or missing Gemini API key (live mode)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Rate limited (this API or the Gemini key's quota)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Assessment",
        "tags": [
          "assessments"
        ]
      }
    },
...
    "/api/v1/assessments/{assessment_id}/logs": {
      "get": {
        "operationId": "list_logs_api_v1_assessments__assessment_id__logs_get",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "maxLength": 36,
              "minLength": 36,
              "title": "Assessment Id",
              "type": "string"
            }
          },
          {
            "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
            "in": "header",
            "name": "x-device-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
              "title": "X-Device-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProgressLogOut"
                  },
                  "title": "Response List Logs Api V1 Assessments  Assessment Id  Logs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List Logs",
        "tags": [
          "progress"
        ]
      },
      "post": {
        "operationId": "add_log_api_v1_assessments__assessment_id__logs_post",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "maxLength": 36,
              "minLength": 36,
              "title": "Assessment Id",
              "type": "string"
            }
          },
          {
            "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
            "in": "header",
            "name": "x-device-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressLogIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressLogOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Add Log",
        "tags": [
          "progress"
        ]
      }
    },
    "/api/v1/assessments/{assessment_id}/logs/{log_id}": {
      "delete": {
        "operationId": "delete_log_api_v1_assessments__assessment_id__logs__log_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "maxLength": 36,
              "minLength": 36,
              "title": "Assessment Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "log_id",
            "required": true,
            "schema": {
              "title": "Log Id",
              "type": "integer"
            }
          },
          {
            "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
            "in": "header",
            "name": "x-device-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
              "title": "X-Device-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Delete Log",
        "tags": [
          "progress"
        ]
      }
    },
    "/api/v1/assessments/{assessment_id}/progress": {
      "get": {
        "operationId": "get_progress_api_v1_assessments__assessment_id__progress_get",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "maxLength": 36,
              "minLength": 36,
              "title": "Assessment Id",
              "type": "string"
            }
          },
          {
            "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
            "in": "header",
            "name": "x-device-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
              "title": "X-Device-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Progress",
        "tags": [
          "progress"
        ]
      }
    },
...
  }
}